mizgirev_TZ_main/requirements.txt
pytest==7.4.3
selenium==4.14.0
retry~=0.9.2
Faker~=23.2.1
allure-pytest==2.13.2
requests==2.28.1

